requirements.txt
numpy==1.16.5
pandas==1.2.5
texar-pytorch==0.1.3
tqdm==4.63.0
torch>=1.9.0
